Resolved Conflicts:
	daemon/src/account.cpp
	daemon/src/dbus/callmanager.h
	daemon/src/iax/iaxaccount.cpp
	daemon/src/sip/sdp.cpp
	daemon/src/sip/sipaccount.cpp
	daemon/src/sip/sipcall.cpp
	daemon/src/sip/sipvoiplink.cpp
	gnome/src/actions.c
	gnome/src/codeclist.c
	gnome/src/config/audioconf.c
	gnome/src/contacts/calltree.c
	kde/plasma/plasmoid-qml/package/contents/ui/BookmarkTab.qml
	kde/plasma/plasmoid-qml/package/contents/ui/CallTab.qml
	kde/plasma/plasmoid-qml/package/contents/ui/ContactTab.qml
	kde/plasma/plasmoid-qml/package/contents/ui/DialPad.qml
	kde/plasma/plasmoid-qml/package/contents/ui/HistoryTab.qml
	kde/plasma/plasmoid-qml/package/contents/ui/NumberOverlay.qml
	kde/plasma/plasmoid-qml/package/contents/ui/Settingstab.qml
	kde/plasma/plasmoid-qml/package/contents/ui/TransferOverlay.qml
	kde/plasma/plasmoid-qml/package/contents/ui/main.qml
	kde/plasma/plasmoid-qml/package/metadata.desktop
	kde/src/AccountListModel.cpp
	kde/src/AccountWizard.cpp
	kde/src/CallView.cpp
	kde/src/SFLPhoneView.cpp
	kde/src/conf/dlgaddressbook.cpp
	kde/src/conf/dlgaudio.cpp
	kde/src/conf/dlggeneral.cpp
	kde/src/klib/AkonadiBackend.cpp
	kde/src/klib/AkonadiBackend.h
	kde/src/klib/CMakeLists.txt
	kde/src/klib/ConfigurationSkeleton.cpp
	kde/src/klib/HelperFunctions.cpp
	kde/src/klib/SortableDockCommon.h
	kde/src/klib/SortableDockCommon.hpp
	kde/src/klib/dataengine/CMakeLists.txt
	kde/src/klib/dataengine/sflphonEngine.cpp
	kde/src/klib/dataengine/sflphonEngine.h
	kde/src/klib/dataengine/sflphone.operations
	kde/src/klib/dataengine/sflphoneService.cpp
	kde/src/klib/dataengine/sflphoneService.h
	kde/src/klib/sflphone-client-kde.kcfg
	kde/src/lib/AccountList.cpp
	kde/src/lib/CallModel.cpp
	kde/src/lib/CallModel.h
	kde/src/lib/CallModel.hpp
	kde/src/lib/sflphone_const.h
	kde/src/widgets/BookmarkDock.cpp
	kde/src/widgets/BookmarkDock.h
	kde/src/widgets/CategorizedTreeWidget.cpp
	kde/src/widgets/CategoryDrawer.h
	kde/src/widgets/ContactDock.cpp
	kde/src/widgets/HistoryDock.cpp
	kde/src/widgets/HistoryTreeItem.cpp
	tools/build-system/launchpad/sflphone-client-kde/debian/rules

+++ daemon/src/account.cpp
@@ -75,6 +75,9 @@ void Account::setRegistrationState(const RegistrationState &state)
 
 void Account::loadDefaultCodecs()
 {
+    // TODO
+    // CodecMap codecMap = Manager::instance ().getCodecDescriptorMap ().getCodecsMap();
+
     // Initialize codec
     std::vector<std::string> result;
     result.push_back("0");
@@ -108,7 +111,7 @@ void Account::setActiveCodecs(const std::vector<std::string> &list)
     for (std::vector<std::string>::const_iterator iter = list.begin(); iter != list.end();
             ++iter) {
         int payload = std::atoi(iter->c_str());
-        audioCodecList_.push_back(static_cast<int>(payload));
+        audioCodecList_.push_back(payload);
     }
 
     // update the codec string according to new codec selection
